+++ internal/services/filesharer/internal/handler/remote_file_test.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"net/http"
 	"net/http/httptest"
+	"os"
 	"testing"
 
 	"github.com/number571/go-peer/pkg/logger"
@@ -66,6 +67,22 @@ func TestHandleRemoteFileAPI(t *testing.T) {
 	if err := remoteFileRequestOK(handlerC); err == nil {
 		t.Fatal("success request with build stream error")
 	}
+
+	tmpFile := "./testdata/hls-filesharer.stg/private/87cb5fb20c1faea4c881c869e2eea4e1b7a20f12d6449efdf1db6255ee5f6b67907d42d06885cc28343cbf62da2d4da9/something.txt"
+	if err := os.WriteFile(tmpFile, []byte("aaaBBBccc"), 0600); err != nil {
+		t.Fatal(err)
+	}
+	defer func() { _ = os.Remove(tmpFile) }()
+
+	if err := remoteFileRequestDeleteOK(handlerA); err != nil {
+		t.Fatal(err)
+	}
+	if err := remoteFileRequestDeleteNotFoundFile(handlerA); err != nil {
+		t.Fatal(err)
+	}
+	if _, err := os.Stat(tmpFile); !errors.Is(err, os.ErrNotExist) {
+		t.Fatal("file is exist after remove")
+	}
 }
 
 func remoteFileRequestOK(handler http.HandlerFunc) error {
@@ -77,6 +94,36 @@ func remoteFileRequestOK(handler http.HandlerFunc) error {
 	defer func() { _ = res.Body.Close() }()
 
 	if res.StatusCode != http.StatusOK {
+		return errors.New("bad status code") // nolint: err113
+	}
+
+	return nil
+}
+
+func remoteFileRequestDeleteOK(handler http.HandlerFunc) error {
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodDelete, "/?friend=abc&name=something.txt", nil)
+
+	handler(w, req)
+	res := w.Result()
+	defer func() { _ = res.Body.Close() }()
+
+	if res.StatusCode != http.StatusOK {
+		return errors.New("bad status code") // nolint: err113
+	}
+
+	return nil
+}
+
+func remoteFileRequestDeleteNotFoundFile(handler http.HandlerFunc) error {
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodDelete, "/?friend=abc&name=111222333_not_exist.txt", nil)
+
+	handler(w, req)
+	res := w.Result()
+	defer func() { _ = res.Body.Close() }()
+
+	if res.StatusCode != http.StatusInternalServerError {
 		return errors.New("bad status code") // nolint: err113
 	}