+++ pkg/api/services/filesharer/client/types.go
@@ -11,7 +11,8 @@ type IClient interface {
 	GetIndex(context.Context) (string, error)
 
 	GetRemoteList(context.Context, string, uint64, bool) (fileinfo.IFileInfoList, error)
-	GetRemoteFile(io.Writer, context.Context, string, string, bool) (bool, string, error)
+	GetRemoteFile(io.Writer, context.Context, string, string, bool) (bool, error)
+	DelRemoteFile(context.Context, string, string) error
 	GetRemoteFileInfo(context.Context, string, string, bool) (fileinfo.IFileInfo, error)
 
 	GetLocalList(context.Context, string, uint64) (fileinfo.IFileInfoList, error)
@@ -25,7 +26,8 @@ type IRequester interface {
 	GetIndex(context.Context) (string, error)
 
 	GetRemoteFileInfo(context.Context, string, string, bool) (fileinfo.IFileInfo, error)
-	GetRemoteFile(io.Writer, context.Context, string, string, bool) (bool, string, error)
+	GetRemoteFile(io.Writer, context.Context, string, string, bool) (bool, error)
+	DelRemoteFile(context.Context, string, string) error
 	GetRemoteList(context.Context, string, uint64, bool) (fileinfo.IFileInfoList, error)
 
 	GetLocalFileInfo(context.Context, string, string) (fileinfo.IFileInfo, error)